feat(ui): extract neomorphic frame styles
